added limited view

+++ app/src/main/java/xmlParser/framework/NodeFinder.java
@@ -6,6 +6,6 @@ import java.awt.*;
 import java.util.Map;
 
 public interface NodeFinder {
-    long findClosestNodeToPoint(int x, int y, Map<Long, CustomNode> nodes, int xOffset, int yOffset, double scaleFactor);
-    Point convertCoordsXYToImageXY(int x, int y, int xOffset, int yOffset, double scaleFactor);
+    long findClosestNodeToPoint(double x, double y, Map<Long, CustomNode> nodes, int xOffset, int yOffset, double scaleFactor);
+    Point convertCoordsXYToImageXY(double x, double y, int xOffset, int yOffset, double scaleFactor);
 }
